+++ bootstrap/Visualisations/bootstrap_twoCat.js
@@ -165,7 +165,7 @@ class bootstrap_twoCat extends visBase {
 
 		// this.largeCISplit = Math.abs(this.populationStatistic - statlist[10000*0.95]);
 		statList.sort(function(a, b){
-			return parseInt(a) - parseInt(b);
+			return parseFloat(a) - parseFloat(b);
 		});
 		let lower_CI = statList[Math.floor(this.numSamples*0.025)];
 		let upper_CI = statList[Math.ceil(this.numSamples*0.975)];
@@ -174,7 +174,7 @@ class bootstrap_twoCat extends visBase {
 	setUpCI(statList){
 		var CISplit = Math.abs(this.populationStatistic - statList[this.numSamples*0.95]);
 		statList.sort(function(a, b){
-			return parseInt(a) - parseInt(b);
+			return parseFloat(a) - parseFloat(b);
 		});
 		let lower_CI = statList[Math.floor(this.numSamples*0.025)];
 		let upper_CI = statList[Math.ceil(this.numSamples*0.975)];
